Modularize Code to make it more readable and easier to work with (#2)
* Improved implementation for odd # of plies

- Still only works for symmetric laminates
- Takes odd # ply laminates and turns them into even # ply laminates by dividing the middle ply in half
- Calculating principal and primary stresses is a WIP
Yet-to-be-done (please bear with me):
- Unit tests
- Break the code into functions that get called in main

* Updated Todo list to reflect recent code commit

* Calculates primary stresses and strains

This is essentially the backbone of the calculator. It covers tasks 6-11 excluding 7. The user inputs laminate stacking sequence, ply thicknesses, Q matrix, and loading condition. Calculator calculates principal and primary stresses and strains. Only symmetric laminates can be analyzed.

* Added Todo items to reflect recent commit

* Updated stipulation "Only works for symmetric ply thicknesses for now"

* Modularize code

Rewrite code so that it is broken up into a number of small files for ease of understandability and for ease of testing.

* Add Tsai_Wu and move constants

Add the Tsai_wu code from the main branch back in move the various constans in Q_Qbar_and_constants into their one file.

* Asthetic Changes

* Minor changes

Remove unnesscicary imports (except matplotlib as that will be used in the future) and fixed spelling and format issues

* Remvoing pycach

* Changing variable name

converting from camel case to snake case

* Merge branch 'Modularize'

* Modularize code

Rewrite code so that it is broken up into a number of small files for ease of understandability and for ease of testing.

* Minor Typos

* Addressing merge conflict

* Fixing typo with np.full